+++ src/modules/index.ts
@@ -13,9 +13,13 @@
  * Registry-based modules (installed via /add-<name> skills, pulled from the
  * `modules` branch): append imports below.
  */
-import './interactive/index.js';
+// Approvals (default tier) must load before self-mod (optional) so the
+// registerApprovalHandler / requestApproval symbols are bound when self-mod
+// registers its handlers at import time.
 import './approvals/index.js';
+import './interactive/index.js';
 import './scheduling/index.js';
 import './permissions/index.js';
 import './agent-to-agent/index.js';
+import './self-mod/index.js';
