+++ src/modules/permissions/permissions.test.ts
@@ -1,15 +1,19 @@
+/**
+ * Tests for the permissions module — canAccessAgentGroup, role helpers, and
+ * ensureUserDm. Moved here from src/access.test.ts in PR #7 alongside the
+ * approvals re-tier that deleted src/access.ts.
+ */
 import { beforeEach, afterEach, describe, expect, it } from 'vitest';
 
-import { pickApprovalDelivery, pickApprover } from './access.js';
-import type { ChannelAdapter, OutboundMessage } from './channels/adapter.js';
-import { initChannelAdapters, registerChannelAdapter, teardownChannelAdapters } from './channels/channel-registry.js';
-import { closeDb, createAgentGroup, createMessagingGroup, initTestDb, runMigrations } from './db/index.js';
-import { canAccessAgentGroup } from './modules/permissions/access.js';
-import { addMember, isMember } from './modules/permissions/db/agent-group-members.js';
-import { createUser } from './modules/permissions/db/users.js';
-import { grantRole, hasAnyOwner, isOwner } from './modules/permissions/db/user-roles.js';
-import { getUserDm } from './modules/permissions/db/user-dms.js';
-import { ensureUserDm } from './modules/permissions/user-dm.js';
+import type { ChannelAdapter, OutboundMessage } from '../../channels/adapter.js';
+import { initChannelAdapters, registerChannelAdapter, teardownChannelAdapters } from '../../channels/channel-registry.js';
+import { closeDb, createAgentGroup, createMessagingGroup, initTestDb, runMigrations } from '../../db/index.js';
+import { canAccessAgentGroup } from './access.js';
+import { addMember, isMember } from './db/agent-group-members.js';
+import { createUser } from './db/users.js';
+import { grantRole, hasAnyOwner, isOwner } from './db/user-roles.js';
+import { getUserDm } from './db/user-dms.js';
+import { ensureUserDm } from './user-dm.js';
 
 function now(): string {
   return new Date().toISOString();
@@ -25,11 +29,6 @@ afterEach(async () => {
   closeDb();
 });
 
-/**
- * Register and activate a mock adapter for tests. `openDM` optional — omit
- * to simulate direct-addressable channels (Telegram/WhatsApp), provide to
- * simulate resolution-required channels (Discord/Slack).
- */
 async function mountMockAdapter(
   channelType: string,
   openDM?: (handle: string) => Promise<string>,
@@ -160,33 +159,9 @@ describe('role helpers', () => {
   });
 });
 
-describe('pickApprover', () => {
-  beforeEach(() => {
-    seedAgentGroup('ag-1');
-    seedAgentGroup('ag-2');
-  });
-
-  it('prefers scoped admins, then globals, then owners — deduplicated', () => {
-    seedUser('u-owner', 'telegram');
-    seedUser('u-ga', 'telegram');
-    seedUser('u-sa', 'telegram');
-    grantRole({ user_id: 'u-owner', role: 'owner', agent_group_id: null, granted_by: null, granted_at: now() });
-    grantRole({ user_id: 'u-ga', role: 'admin', agent_group_id: null, granted_by: null, granted_at: now() });
-    grantRole({ user_id: 'u-sa', role: 'admin', agent_group_id: 'ag-1', granted_by: null, granted_at: now() });
-
-    expect(pickApprover('ag-1')).toEqual(['u-sa', 'u-ga', 'u-owner']);
-    expect(pickApprover('ag-2')).toEqual(['u-ga', 'u-owner']);
-    expect(pickApprover(null)).toEqual(['u-ga', 'u-owner']);
-  });
-
-  it('returns empty list when nobody is privileged', () => {
-    expect(pickApprover('ag-1')).toEqual([]);
-  });
-});
-
 describe('ensureUserDm', () => {
   it('adapter without openDM: falls through to using the bare handle as platform_id', async () => {
-    await mountMockAdapter('nodm'); // no openDM → direct-addressable fallback
+    await mountMockAdapter('nodm');
     seedUser('nodm:123', 'nodm');
 
     const mg = await ensureUserDm('nodm:123');
@@ -195,16 +170,11 @@ describe('ensureUserDm', () => {
     expect(mg!.platform_id).toBe('123');
     expect(mg!.is_group).toBe(0);
 
-    // Cache row written
     const cached = getUserDm('nodm:123', 'nodm');
     expect(cached?.messaging_group_id).toBe(mg!.id);
   });
 
   it('Telegram via chat-sdk-bridge: adapter.openDM returns prefixed platform_id', async () => {
-    // Post-fix bridge behavior: the bridged Telegram adapter exposes openDM
-    // that delegates to the underlying @chat-adapter/telegram adapter, whose
-    // channelIdFromThreadId returns "telegram:<chatId>". That's the same
-    // encoding onInbound stores in messaging_groups, so cache hits on repeat.
     const mock = await mountMockAdapter('telegram', async (handle) => `telegram:${handle}`);
     seedUser('telegram:6037840640', 'telegram');
 
@@ -213,7 +183,6 @@ describe('ensureUserDm', () => {
     expect(mg!.platform_id).toBe('telegram:6037840640');
     expect(mock.openDMCalls).toEqual(['6037840640']);
 
-    // Second call hits the user_dms cache, not openDM again.
     const mg2 = await ensureUserDm('telegram:6037840640');
     expect(mg2!.id).toBe(mg!.id);
     expect(mock.openDMCalls).toEqual(['6037840640']);
@@ -228,10 +197,9 @@ describe('ensureUserDm', () => {
     expect(mg!.platform_id).toBe('dm-channel-user-1');
     expect(mock.openDMCalls).toEqual(['user-1']);
 
-    // Second call should hit the cache, not openDM.
     const mg2 = await ensureUserDm('discord:user-1');
     expect(mg2!.id).toBe(mg!.id);
-    expect(mock.openDMCalls).toEqual(['user-1']); // unchanged
+    expect(mock.openDMCalls).toEqual(['user-1']);
   });
 
   it('returns null when the adapter is not registered', async () => {
@@ -245,7 +213,6 @@ describe('ensureUserDm', () => {
     });
     seedUser('slack:u1', 'slack');
     expect(await ensureUserDm('slack:u1')).toBeNull();
-    // No cache row should be written on failure
     expect(getUserDm('slack:u1', 'slack')).toBeUndefined();
   });
 
@@ -266,47 +233,5 @@ describe('ensureUserDm', () => {
     const mg = await ensureUserDm('telegram:555');
     expect(mg?.id).toBe('mg-preexisting');
     expect(getUserDm('telegram:555', 'telegram')?.messaging_group_id).toBe('mg-preexisting');
-  });
-});
-
-describe('pickApprovalDelivery', () => {
-  beforeEach(() => {
-    seedAgentGroup('ag-1');
-  });
-
-  it('returns the first reachable approver', async () => {
-    await mountMockAdapter('telegram');
-    seedUser('telegram:111', 'telegram');
-    seedUser('telegram:222', 'telegram');
-
-    // Both users are reachable (direct-addressable), so the first wins.
-    const result = await pickApprovalDelivery(['telegram:111', 'telegram:222'], 'telegram');
-    expect(result?.userId).toBe('telegram:111');
-    expect(result?.messagingGroup.platform_id).toBe('111');
-  });
-
-  it('prefers same-channel-kind approver on tie-break', async () => {
-    await mountMockAdapter('telegram');
-    await mountMockAdapter('discord', async (h) => `dm-${h}`);
-    seedUser('telegram:111', 'telegram');
-    seedUser('discord:222', 'discord');
-
-    // Origin is discord → discord approver wins even though telegram is first.
-    const result = await pickApprovalDelivery(['telegram:111', 'discord:222'], 'discord');
-    expect(result?.userId).toBe('discord:222');
-  });
-
-  it('falls through to any reachable approver when none match origin', async () => {
-    await mountMockAdapter('telegram');
-    seedUser('telegram:111', 'telegram');
-
-    const result = await pickApprovalDelivery(['telegram:111'], 'discord');
-    expect(result?.userId).toBe('telegram:111');
-  });
-
-  it('returns null when nobody is reachable', async () => {
-    // No adapter registered → no user is reachable.
-    seedUser('telegram:111', 'telegram');
-    expect(await pickApprovalDelivery(['telegram:111'], 'telegram')).toBeNull();
   });
 });
